% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_clean.R
\name{cleanhydro}
\alias{cleanhydro}
\title{Clean raw hydrologic DBHYDRO data retrievals}
\usage{
cleanhydro(res)
}
\arguments{
\item{dt}{data.frame output of gethydro}
}
\description{
Cleans to by consistent with water quality formatting. Connects metadata header to actual measurements.
}
\examples{
\dontrun{
cleanhydro(gethydro(dbkey="15081",date_min="2013-01-01",date_max="2013-02-02"))
}
}

