% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_clean.R
\name{cleanwq}
\alias{cleanwq}
\title{Clean raw water quality DBHYDRO data retrievals}
\usage{
cleanwq(dt)
}
\arguments{
\item{dt}{data.frame output of getwq}
}
\description{
remove extra qa/db columns, remove qa "blanks", and convert results from long to wide format.
}
\examples{
dt<-read.csv(system.file("extdata","testwq.csv",package="dbhydroR"))
cleanwq(dt)
}

