% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_clean.R
\name{dbhydro_clean}
\alias{dbhydro_clean}
\title{Clean raw DBHYDRO data retrievals}
\usage{
dbhydro_clean(dt)
}
\arguments{
\item{dt}{data.frame output of dbhydro_get}
}
\description{
remove extra qa/db columns, remove qa "blanks", and convert results from long to wide format.
}
\examples{
dt<-read.csv(system.file("extdata","test.csv",package="dbhydroR"))
dbhydro_clean(dt)
}

