% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydoR-package.R
\docType{package}
\name{dbhydroR-package}
\alias{dbhydroR}
\alias{dbhydroR-package}
\title{dbhydroR}
\description{
dbhydroR is an R interace to the South Florida Water Management District's DBHYDRO database which holds over 35 million hydrologic and water quality records from the Florida Everglades and surrounding areas.
}
\section{Hydrologic Data}{

The \code{\link[dbhydroR]{get_hydro}} function provides the capability to return:
\itemize{
  \item weather data
  \item surfacewater data
  \item groundwater data
  \item water-quality sonde data
}
}

\section{Water Quality Data}{

The \code{\link[dbhydroR]{get_wq}} function provides the capability to return:
\itemize{
  \item water quality data
}
}
\keyword{package}

