% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_plot.R
\name{dbydro_plot}
\alias{dbhydro_plot}
\alias{dbydro_plot}
\title{Plot DBHYDRO data retrievals}
\usage{
dbhydro_plot(dt, label = "topleft", type = "b", abb = Inf)
}
\arguments{
\item{dt}{data.frame output of dbhydro_get}

\item{label}{string placement of site label choice of "topleft", "topright", "bottomleft", or "bottomright"}

\item{type}{string plot type}

\item{abb}{integer number of characters to abbreviate y-axis label}
}
\description{
Plot DBHYDRO data retrievals
}
\examples{
dt<-dbhydro_get(station_id=c("FLAB08","FLAB09"),date_min="2011-03-01",date_max="2014-05-01",test_name=c("CHLOROPHYLLA-SALINE","SALINITY"))
dbhydro_plot(dt,label="bottomleft",abb=3)
}

