% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get.R
\name{get_dbkey}
\alias{get_dbkey}
\alias{getdbkey}
\title{Query dbkey information}
\usage{
get_dbkey(
  category,
  stationid = NA,
  param = NA,
  freq = NA,
  longest = FALSE,
  stat = NA,
  recorder = NA,
  agency = NA,
  strata = NA,
  detail.level = "summary",
  ...
)
}
\arguments{
\item{category}{character string, choice of "WEATHER", "SW", "GW", or "WQ"}

\item{stationid}{character string specifying station name}

\item{param}{character string specifying desired parameter name}

\item{freq}{character string specifying collection frequency (daily = "DA")}

\item{longest}{logical limit results to the longest period-of-record?}

\item{stat}{character string specifying statistic type}

\item{recorder}{character string specifying recorder information}

\item{agency}{character string specifying collector agency}

\item{strata}{numeric vector of length 2 specifying a range of z-coordinates
relative to local ground elevation. Only applicable for queries in the
"WEATHER" and "GW" categories.}

\item{detail.level}{character string specifying the level of detail to return.
Choices are "full", "summary", and "dbkey".}

\item{...}{Options passed as named parameters}
}
\description{
Retrieve a data.frame summary including dbkeys or a vector of
dbkeys corresponding to specified parameters
}
\details{
A \code{dbkey} represents a unique station x variable time-series. A
value in the "Recorder" field of "PREF" should be used whenever possible.
This indicates that the dataset has been checked by the SFWMD modelling
group.
}
\examples{
\dontrun{
# Weather
get_dbkey(stationid = "JBTS", category = "WEATHER", param = "WNDS",
detail.level = "summary")
get_dbkey(stationid = "JBTS", category = "WEATHER", param = "WNDS",
detail.level = "dbkey")

# query on multiple values
get_dbkey(stationid = c("MBTS", "JBTS"), category = "WEATHER",
param = "WNDS", freq = "DA", detail.level = "dbkey")


# Surfacewater
get_dbkey(stationid = "C111\%", category = "SW")
get_dbkey(category = "SW", stationid = "LAKE\%", detail.level = "full")

# Groundwater
get_dbkey(stationid = "C111\%", category = "GW")
get_dbkey(stationid = "C111AE", category = "GW", param = "WELL",
freq = "DA", stat = "MEAN", strata = c(9, 22), recorder = "TROL",
 agency = "WMD", detail.level = "full")

# Water Quality
get_dbkey(stationid = "C111\%", category = "WQ")
}
}
