% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_get.R
\name{getdbkey}
\alias{getdbkey}
\title{Retrive a list of dbkeys from a DBHYDRO station ID}
\usage{
getdbkey(stationid, category, param = NA, freq = "DA", blind = FALSE)
}
\arguments{
\item{stationid}{character string}

\item{category}{character string, choice of "WEATHER","SW","GW", or "WQ"}

\item{blind}{logical output dbkey results as object (TRUE) or simply print query results (FALSE)?}

\item{string}{desired parameter name}
}
\description{
Retrive a list of dbkeys from a DBHYDRO station ID
}
\examples{
getdbkey(stationid="JBTS",category="WEATHER",param="WNDS")
getdbkey(stationid="JBTS",category="WEATHER",param="WNDS",blind=TRUE)
getdbkey(stationid="C111\%",category="SW")
getdbkey(stationid="C111\%",category="GW")
getdbkey(stationid="C111\%",category="WQ")
}

