% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_get.R
\name{gethydro}
\alias{gethydro}
\title{Retrieve hydrologic data from the DBHYDRO Environmental Database}
\usage{
gethydro(dbkey, date_min = NA, date_max = NA, period = "uspec",
  raw = FALSE, v_target_code = "file_csv")
}
\arguments{
\item{dbkey}{character string of time series identifiers (e.g. Joe Bay mean daily wind speed is "15081"). These are listed alongside each time series in DBHYDRO.}

\item{date_min}{character must be in POSIXct YYYY-MM-DD format}

\item{date_max}{character must be in POSIXct YYYY-MM-DD format}
}
\description{
South Florida Water Management District
}
\examples{
#One variable/station time series
gethydro(dbkey="15081",date_min="2013-01-01",date_max="2013-02-02")

#Multiple variable/station time series
gethydro(dbkey=c("15081","15069"),date_min="2013-01-01",date_max="2013-02-02")
}

