% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbhydro_plot.R
\name{plotdbydro}
\alias{plotdbhydro}
\alias{plotdbydro}
\title{Plot DBHYDRO data retrievals}
\usage{
plotdbhydro(dt, label = "topleft", type = "b", abb = Inf)
}
\arguments{
\item{dt}{data.frame output of dbhydro_get}

\item{label}{string placement of site label choice of "topleft", "topright", "bottomleft", or "bottomright"}

\item{type}{string plot type}

\item{abb}{integer number of characters to abbreviate y-axis label}
}
\description{
Plot DBHYDRO data retrievals
}
\examples{
\dontrun{
dt<-getwq(station_id=c("FLAB08","FLAB09"),
date_min="2011-03-01",date_max="2014-05-01",
test_name=c("CHLOROPHYLLA-SALINE","SALINITY"))
plotdbhydro(dt,label="bottomleft",abb=3)

dt<-gethydro(stationid="JBTS",category="WEATHER",param="WNDS",date_min="2011-03-01",
date_max="2012-05-01")

plotdbhydro(dt,label="bottomleft",abb=3)
}
}

