% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_clean.R
\name{cleanwq}
\alias{cleanwq}
\title{Clean raw water quality DBHYDRO data retrievals}
\usage{
cleanwq(dt)
}
\arguments{
\item{dt}{data.frame output of getwq}
}
\description{
Remove extra columns associated with QA flags, LIMS, and District recieving. Remove QA "blanks". Convert results from long to wide format.
}
\details{
Current DBHYDRO practice is to return values below the MDL as 0 minus the uncertainty estimate.
}
\examples{
\dontrun{
#check handling of values below MDL
dt <- getwq(station_id = "FLAB08", date_min = "2009-01-01",
date_max = "2016-02-29", test_name = "PHOSPHATE, TOTAL AS P", raw = TRUE)
dt <- getwq(station_id = "FLAB08", date_min = "2009-01-01",
date_max = "2016-02-29", test_name = "AMMONIA-N", raw = TRUE)
}

dt <- read.csv(system.file("extdata", "testwq.csv", package = "dbhydroR"))
cleanwq(dt)
}

