% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get.R
\name{getdbkey}
\alias{getdbkey}
\title{Retrieve a list of dbkeys from a DBHYDRO station ID}
\usage{
getdbkey(stationid, category, param = NA, freq = "DA", blind = FALSE)
}
\arguments{
\item{stationid}{character string}

\item{category}{character string, choice of "WEATHER","SW","GW", or "WQ"}

\item{param}{string optional desired parameter name}

\item{freq}{character string frequency choice of daily ("DA")}

\item{blind}{logical output dbkey results as object (TRUE) or simply print query results (FALSE)?}
}
\details{
A value in the "Recorder" field on "PREF" should be used whenever possible. This indicates that the dataset has been checked by the SFWMD modeling group.
}
\examples{
\dontrun{
getdbkey(stationid="JBTS",category="WEATHER",param="WNDS")
getdbkey(stationid="C111\%",category="SW")
getdbkey(stationid="C111\%",category="GW")
getdbkey(stationid="C111\%",category="WQ")
getdbkey(stationid="JBTS",category="WEATHER",param="WNDS")
}
}

